:truck: 3.6.3: Rename the addons.
